system8/requirements.txt
# RAG評価フレームワーク - Requirements

# LlamaIndex関連
llama-index==0.9.13
llama-index-core==0.9.13
llama-index-llms-ollama==0.1.1
llama-index-embeddings-ollama==0.1.1
llama-index-embeddings-openai==0.1.5
llama-index-embeddings-huggingface==0.1.4
llama-index-vector-stores-milvus==0.1.5
llama-index-vector-stores-qdrant==0.1.3
llama-index-vector-stores-chroma==0.1.4
llama-index-readers-file==0.1.4
llama-index-storage-docstore-mongodb==0.1.2
llama-index-storage-kvstore-redis==0.1.3

# 埋め込みモデル
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
accelerate==0.24.1

# ベクターストア
chromadb==0.4.15
qdrant-client==1.6.9
pymilvus==2.3.3

# 評価フレームワーク
ragas==0.0.22
datasets==2.14.7
rouge-score==0.1.2
bert-score==0.3.13
evaluate==0.4.1

# データ処理
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# 可視化
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# 設定・ユーティリティ
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# 開発・テスト
pytest==7.4.3
pytest-asyncio==0.21.1
jupyter==1.0.0
ipykernel==6.26.0

# PDF処理
pypdf==3.17.1
pymupdf==1.23.8

# 自然言語処理
spacy==3.7.2
nltk==3.8.1
mecab-python3==1.0.6

# 統計処理
scipy==1.11.4
statsmodels==0.14.0

# OpenAI (オプション)
openai==1.3.7
tiktoken==0.5.2

# その他のLLM API
anthropic==0.7.8

# データベース (オプション)
redis==5.0.1
pymongo==4.6.0
neo4j==5.15.0

# 並列処理
joblib==1.3.2
asyncio

# 型ヒント
typing-extensions==4.8.0

# ログ・監視
psutil==5.9.6